Improve copy/rename of states with nesting content
When a state is copy-pasted, and then renamed, the tool now checks if
the copied state was nested. If so it makes a deep copy of the nested
state at the time when it is renamed. This way the nested content is not
lost and does not refer to the same nested state.