Once in the project folder, don't move out
When loading a model in a different folder, the tool has to move to this
folder to find the ASN.1 files that are referenced without any path
information.
The tool used to return immediately to the folder where is was called
from. The problem is that subsequent model checks would fail because
they would not know the path of the asn1 files anymore.