Major refactoring to move the context to ASN.1
Now all the context, including the state and all DCL variables are
generated as an ASN.1 type
This allows a better interface with external tools in C or Python since
there is a single ASN.1 SEQUENCE variable. There is no need to have a
getter and a setter for each internal variable.